wo #5532 . refactor lp46 for use in lp56
